requirements.txt
python-dotenv==0.20.0
SQLAlchemy==1.3.24
gino==1.0.1
aiogram==2.21
asyncpg==0.26.0
aiosignal==1.3.1
async-timeout==4.0.2
aiohttp==3.8.3
asyncpg==0.26.0
attrs==22.1.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.3
pytz==2022.6
setuptools==65.5.0
wheel==0.37.1
yarl==1.8.2